_clearForm method
``_clearForm`` method has been created to handle the removing value from the form effieciently. It will remove the value if the element is input,textarea, select but for checkbox and radio button.  radio button and checkbox should be handled differently it applies checked false.

I have add *noEmpty* class exclusion for handy to user. We can specifically handle a particular field using this class. when we apply this class to form element, cloning will not remove the form element value. (may be need documention)

First clone form should remain, but it values should be flushed when we click on delete. I did that in this commit