preserveChildCount should not require cloned data
Now that cloning of data values is fixed, it's not safe to rely on the
inNewClone data value to get the initial child count as this may not
be present.

Instead, get the value from the original child in the toClone element
and copy it to the relevant element in the newClone manually.